Remove RocksDB remnants, cleanup
