+++ src/bin/db_inspector.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use locustdb::disk_store::v2::{Storage, StorageV2};
+use locustdb::disk_store::storage::Storage;
 use locustdb::perf_counter::PerfCounter;
 use structopt::StructOpt;
 
@@ -12,9 +12,9 @@ use structopt::StructOpt;
     author = "Clemens Winter <clemenswinter1@gmail.com>"
 )]
 struct Opt {
-    /// Path to data directory based on v2 storage format
-    #[structopt(long, name = "PATH_V2", parse(from_os_str))]
-    db_v2_path: PathBuf,
+    /// Database path
+    #[structopt(long, name = "PATH", parse(from_os_str))]
+    db_path: PathBuf,
 
     /// Wal detail. 0 = no detail, 1 = number of segments, 2 = tables per segment + total rows, 3 = rows per table, 4 = full table dump
     #[structopt(long, default_value = "1")]
@@ -33,7 +33,7 @@ struct Opt {
 async fn main() {
     env_logger::init();
     let opts = Opt::from_args();
-    let (storage, wal) = StorageV2::new(&opts.db_v2_path, Arc::new(PerfCounter::default()), true);
+    let (storage, wal) = Storage::new(&opts.db_path, Arc::new(PerfCounter::default()), true);
 
     {
         let meta = storage.meta_store().lock().unwrap();
